btrfs: add per subvolume security context
The Docker SELinux support depends on assigning a unique
security context to each container's filesystem using a context=
mount option.  This approach works with the device mapper
backend.  However, it breaks when using btrfs subvolumes
because there is no way to assign a per-subvolume context mount,
and the Docker btrfs backend uses subvolume snapshots for the
container filesystems.  Supporting per-subvolume context mounts
within SELinux is impossible because subvolumes are not visible
to the vfs or SELinux.

Implement per-subvolume context mount support within btrfs
by adding a new mount option, subvol_context=, that can be used
to specify the desired context value when a subvol= mount is
performed.  We cannot reuse the existing context= mount option
because it would be interpreted and consumed by the vfs and
SELinux and would never be seen by btrfs at all.

The implementation works by storing the subvolume context in the
btrfs_root for the subvolume, and returning this context if set
from btrfs_getxattr for requests on security.selinux.  All
inodes within each subvolume are thereby treated as having the
specified context, and different subvolumes may use a different
context.  As with context= mounts, we also prevent userspace
from setting contexts on inodes when subvol_context= was used and we
disable automatic setting of contexts on newly allocated inodes.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>
Signed-off-by: Vincent Batts <vbatts@redhat.com>